requirements.txt
Flask==3.0.0
gunicorn==21.2.0
openai==1.3.7
python-dotenv==1.0.0
twilio==8.9.1